+++ Sources/Enums/Utils.swift
@@ -7,21 +7,69 @@ import Foundation
 
 extension Array {
   /// Returns the element at the specified index iff it is within bounds, otherwise nil.
-  subscript(safe index: Int ) -> Element? {
+  subscript(safe index: Int) -> Element? {
     return indices.contains(index) ? self[index] : nil  /// Returns the element at the specified index iff it is within bounds, otherwise nil.
   }
 }
 
-extension String {
-  func toDouble() -> Double? {
-    return Double(self)
+extension Array where Element: Node {
+  
+  // MARK: String
+  func toString(_ index: Int) -> String? {
+    if let node = self[safe: index] as? VariableNode {
+      return node.name
+    }
+    return nil
   }
-  func toFloat() -> Float? {
-    return Float(self)
+  
+  func contains(_ name: String) -> Bool {
+    for node in self {
+      if node.name == name {
+        return true
+      }
+    }
+    return false
   }
-  func toInt() -> Int? {
-    return Int(self)
+  
+  // MARK: Number
+  func toDouble(_ index: Int) -> Double? {
+    if let node = self[safe: index] as? NumberNode {
+      return node.value
+    }
+    return nil
   }
+  func toDoubles() -> [Double] {
+    let nodes = self.compactMap { $0 as? NumberNode }
+    return nodes.map { Double($0.value) }
+  }
+  func toInt(_ index: Int) -> Int? {
+    if let node = self[safe: index] as? NumberNode {
+      return Int(node.value)
+    }
+    return nil
+  }
+  func toFloat(_ index: Int) -> Float? {
+    if let node = self[safe: index] as? NumberNode {
+      return Float(node.value)
+    }
+    return nil
+  }
+  func toCGFloat(_ index: Int) -> CGFloat? {
+    if let node = self[safe: index] as? NumberNode {
+      return CGFloat(node.value)
+    }
+    return nil
+  }
+  func toBool(_ index: Int) -> Bool? {
+    if let node = self[safe: index] as? NumberNode {
+      return node.value != 0
+    }
+    if let node = self[safe: index] as? VariableNode, let value = Bool(node.name) {
+      return value
+    }
+    return nil
+  }
+  
 }
 
 extension RawRepresentable {
@@ -65,4 +113,55 @@ extension CALayer {
     var currentMediaTime: CFTimeInterval {
        return convertTime(CACurrentMediaTime(), from: nil)
     }
+}
+
+extension String {
+  
+  func parseNameAndParams() -> [Node] {
+    let parser = Parser(tokens: Lexer.tokenize(self.lowercased()))
+    do {
+      return try parser.parse()
+    } catch {
+      #if DEBUG
+      print("Failed to parse \"\(self)\", error: \(error)")
+      #endif
+      return []
+    }
+  }
+  
+  /**
+   Helper function that returns a tuple containing the name and params from a string
+   
+   - Parameter from string: The string to be converted into `enum`.
+   - Discussion: the string format is like "enumName(param1,param2,param3)"
+   - Returns: A tuple containing the name and an array of parameter string
+   */
+  func extractNameAndParams() -> (String, [Node])? {
+    let nodes = self.parseNameAndParams()
+    guard let firstNode = nodes.first else {
+      return nil
+    }
+    let params: [Node] = (firstNode as? CallNode)?.arguments ?? []
+    return (firstNode.name, params)
+  }
+  
+  public func parse<T: IBEnum>() -> T? {
+    return T(string: self)
+  }
+  
+  public func parse<T: IBEnum>(default defaultValue: T) -> T {
+    return T(string: self) ?? defaultValue
+  }
+  
+}
+
+extension NSRegularExpression {
+  
+  func matched(_ string: String) -> (String, CountableRange<Int>)?  {
+    let range = self.rangeOfFirstMatch(in: string, options: [], range: NSRange(0 ..< string.utf16.count))
+    if range.location != NSNotFound {
+      return ((string as NSString).substring(with: range), range.location ..< range.location + range.length)
+    }
+    return nil
+  }
 }